+++ Config/routes.php
@@ -52,7 +52,20 @@ return array(
         'controller' => 'Admin:Form@tweakAction',
         'disallow' => array('guest')
     ),
-    
+
+    // Submit logger
+    '/%s/module/mail-form/submit-logs' => array(
+        'controller' => 'Admin:SubmitLog@indexAction'
+    ),
+
+    '/%s/module/mail-form/submit-logs/view/(:var)' => array(
+        'controller' => 'Admin:SubmitLog@viewAction'
+    ),
+
+    '/%s/module/mail-form/submit-logs/delete/(:var)' => array(
+        'controller' => 'Admin:SubmitLog@deleteAction'
+    ),
+
     // Dynamic fields
     '/%s/module/mail-form/field/add/(:var)' => array(
         'controller' => 'Admin:Field@addAction'